requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Earth Observation
rasterio>=1.3.0
geopandas>=0.12.0
earthpy>=0.9.4

# Image Processing
opencv-python>=4.7.0
albumentations>=1.3.0
Pillow>=9.5.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
scikit-image>=0.20.0

# Metrics & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.12.0
tqdm>=4.65.0

# Model Compression
torch-pruning>=1.2.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
h5py>=3.8.0
jupyterlab>=3.6.0

# Deployment
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6

# Optional (HPC)
# mpi4py>=3.1.4  # Uncomment for multi-node training
